separated client from server code, split tests based on features and improved tests.
