+++ client/static/developer.jpg
