+++ client/templates/logout.html
