+++ client/templates/query.html
