--- nemoApp/__init__.py
@@ -1,8 +1,0 @@
-# imports
-from flask import Flask
-
-app = Flask(__name__)
-app.config.from_object('config')
-
-# avoiding circular reference error, placing import below.
-from nemoApp import views
