+++ server/requirements.txt
@@ -1,14 +1,11 @@
 Flask==0.10.1
 Flask-MySQLdb==0.2.0
-Flask-SQLAlchemy==2.1
 Flask-WTF==0.12
-gunicorn==19.4.5
 itsdangerous==0.24
 Jinja2==2.8
 MarkupSafe==0.23
 mysqlclient==1.3.7
 nose==1.3.7
-SQLAlchemy==1.0.12
 Werkzeug==0.11.4
 wheel==0.29.0
 WTForms==2.1
