+++ server/run.py
@@ -1,6 +1,6 @@
 #!usr/bin/python
 
-from nemoApp import app
+from app import app
 import os  
 port = int(os.environ.get('PORT', 5000))
 app.run(host='0.0.0.0', port=port)
