+++ server/tests/tests_connection.py
@@ -2,7 +2,7 @@
 import unittest, os, datetime
 from flask import json, request, Response, session
 import MySQLdb
-from nemoApp import app
+from app import app
 
 class BasicConnectionTestCase(unittest.TestCase):
     """Checks for app and db connectivity"""
@@ -232,120 +232,6 @@ class BasicConnectionTestCase(unittest.TestCase):
         except MySQLdb.OperationalError as e:# 
             data = None
         self.assertNotEqual(data, None)
-
-class SavingNemoTestCase(unittest.TestCase):
-    """SavingNemo Feature specific Test Cases will go here"""
-
-    def setUp(self):
-        """Setup test app"""
-        app.config['TESTING'] = True
-
-    def tearDown(self):
-        """Destroy test app"""
-        # ...
-
-    def login(self, c, username, password):
-        """Login helper function"""
-        return c.post('/login', data=dict(
-            username=username,
-            password=password,
-            ), follow_redirects=True)
-
-    def logout(self, c):
-        """Logout helper function"""
-        return c.get('/logout', follow_redirects=True)
-
-    def test_login_logout_good(self):
-        """Test Login and Logout using helper functions"""      
-        with app.test_client() as c:
-            rv = self.login(c,
-                app.config['USERNAME'],
-                app.config['PASSWORD'])
-            # Check log in successful
-            self.assertEqual(request.path, '/query')
-            # Check logged in session variable is set to True
-            self.assertTrue(session.get('logged_in'))
-
-            rv = self.logout(c)
-            # Check log out successful
-            self.assertEqual(request.path, '/logout')
-            # Check logged in session variable is set to None
-            self.assertEqual(session.get('logged_in'), None)
-
-    def test_login_invalid_username(self):
-        """Test login with Invalid Username"""
-        with app.test_client() as c:
-            # Check for Invalid Username
-            rv = self.login(c,
-                app.config['USERNAME'] + 'x',
-                app.config['PASSWORD'])
-
-            # Check log in fails
-            self.assertEqual(request.path, '/login')
-            # Check logged in session variable is set to None
-            self.assertEqual(session.get('logged_in'), None)
-            # Check if error message is displayed for invalid username
-            self.assertIn(b"Invalid Username. Please try again.", rv.data)
-
-    def test_login_invalid_password(self):
-        """Test login with Invalid Password"""
-        with app.test_client() as c:
-            # Check for Invalid Password
-            rv = self.login(c,
-                app.config['USERNAME'],
-                app.config['PASSWORD'] + 'x')
-            # Check log in fails
-            self.assertEqual(request.path, '/login')
-            # Check logged in session variable is set to None
-            self.assertEqual(session.get('logged_in'), None)
-            # Check if error message is displayed for invalid password
-            self.assertIn(b"Invalid Password. Please try again.", rv.data)
-
-    def test_query_submit_good_values(self):
-        """Test the QueryForm with proper values"""
-        # Check Fields without redirection
-        with app.test_client() as c:            
-            rv = c.post('/query', 
-                data=dict(
-                    logger_type='Mussel',
-                    country_name='USA',
-                    state_name='Massachusetts', 
-                    date_pick_from=datetime.date(1989, 1, 1).strftime('%m/%d/%Y'),
-                    date_pick_to=datetime.date(2001, 4, 1).strftime('%m/%d/%Y')),
-                follow_redirects=False)
-            self.assertEqual(request.form.get('logger_type'), "Mussel")
-            self.assertEqual(request.form.get('country_name'), "USA")
-            self.assertEqual(request.form.get('state_name'), "Massachusetts")
-            self.assertEqual(request.form.get('date_pick_from'), "01/01/1989")
-            self.assertEqual(request.form.get('date_pick_to'), "04/01/2001")
-
-    def test_query_submit_redirect_success(self):
-        """Test redirection of QueryForm with proper values"""
-        # Check successful redirection
-        with app.test_client() as c:            
-            rv = c.post('/query', 
-                data=dict(
-                    logger_type='Mussel',
-                    country_name='USA',
-                    state_name='Massachusetts', 
-                    date_pick_from=datetime.date(1989, 1, 1).strftime('%m/%d/%Y'),
-                    date_pick_to=datetime.date(2001, 4, 1).strftime('%m/%d/%Y')),
-                follow_redirects=True)
-            # Check successful redirection
-            self.assertEqual(request.path, '/query')
-
-    def test_query_submit_missing_compulsory_values(self):
-        """Test the QueryForm with compulsory fields missing"""
-        with app.test_client() as c:
-            rv = c.post('/query', 
-                data=dict( 
-                    country_name='USA',
-                    state_name='Massachusetts'), 
-                follow_redirects=True)
-            # Check Query Submission fails
-            self.assertEqual(request.path, '/query')
-            # Check if error message is displayed for invalid query
-            self.assertIn(b'Invalid Submission. All fields marked with * are compulsory', rv.data)
 
 if __name__ == '__main__':
     unittest.main()
